+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Add new scatter plot markers: `"|"`, and `"-"`
 * Add drag-and-drop functionality (`drag_drop_source`, `drag_drop_target`)
 * Add `TF.hover` attribute for more robust hover functionality than `Image.hover_tag`.
+* Re-introduce the `maximizied` window flag, because just setting window size to a large number doesn't properly work on Windows
 
 **Breaking Changes**