+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 
-## v0.10.0 (planned)
+## v0.10.0
 
 * Add new scatter plot markers: `"|"`, and `"-"`
 * Add drag-and-drop functionality (`drag_drop_source`, `drag_drop_target`)